Move geog_data fixture to conftest
Use it to reduce duplication with geospatial_gradient test.
