+++ tests/calc/test_kinematics.py
@@ -4,7 +4,6 @@
 """Test the `kinematics` module."""
 
 import numpy as np
-import pyproj
 import pytest
 import xarray as xr
 
@@ -64,24 +63,6 @@ def test_vorticity():
     v = vorticity(u.T, u.T, dx=1 * units.meter, dy=1 * units.meter)
     true_v = np.ones_like(u) / units.sec
     assert_array_equal(v, true_v)
-
-
-@pytest.fixture
-def geog_data(request):
-    """Create data to use for testing calculations on geographic coordinates."""
-    # Generate a field of u and v on a lat/lon grid
-    crs = pyproj.CRS(request.param)
-    proj = pyproj.Proj(crs)
-    a = np.arange(4)[None, :]
-    arr = np.r_[a, a, a] * units('m/s')
-    lons = np.array([-100, -90, -80, -70]) * units.degree
-    lats = np.array([45, 55, 65]) * units.degree
-    lon_arr, lat_arr = np.meshgrid(lons.m_as('degree'), lats.m_as('degree'))
-    factors = proj.get_factors(lon_arr, lat_arr)
-
-    return (crs, lons, lats, arr, arr, factors.parallel_scale, factors.meridional_scale,
-            lat_lon_grid_deltas(lons.m, np.zeros_like(lons.m), geod=crs.get_geod())[0][0],
-            lat_lon_grid_deltas(np.zeros_like(lats.m), lats.m, geod=crs.get_geod())[1][:, 0])
 
 
 @pytest.mark.parametrize('geog_data', ('+proj=lcc lat_1=25', '+proj=latlon', '+proj=stere'),
